+++ sodalite/core/config.py
@@ -12,11 +12,11 @@ logger = logging.getLogger(__name__)
 KEY_HOOKS = 'hooks'
 KEY_KEYMAP = 'keymap'
 KEY_KEYMAP_GLOBAL = 'global'
-KEY_KEYMAP_NORMAL = 'normal'
+KEY_KEYMAP_NAVIGATE = 'navigate'
 KEY_KEYMAP_ASSIGN = 'assign'
 KEY_KEYMAP_EDIT = 'edit'
 KEY_KEYMAP_MODES = (KEY_KEYMAP_GLOBAL,
-                    KEY_KEYMAP_NORMAL,
+                    KEY_KEYMAP_NAVIGATE,
                     KEY_KEYMAP_ASSIGN,
                     KEY_KEYMAP_EDIT)
 PREFERRED_NAMES = 'preferred_names'