+++ sodalite/ui/control.py
@@ -29,7 +29,7 @@ class Control:
         self.ACTION_TO_CALL: Dict[Action, Callable[[]]] = {
             GlobalAction.exit: self.exit,
             GlobalAction.abort: self.abort,
-            GlobalAction.normal_mode: self.normal_mode,
+            GlobalAction.navigate_mode: self.enter_navigate_mode,
             GlobalAction.filter: self.trigger_filter,
             GlobalAction.yank_current_path: self.yank_to_clipboard,
             GlobalAction.toggle_dotfiles: self.toggle_dotfiles,
@@ -71,8 +71,8 @@ class Control:
     def abort(self):
         graphics.exit()
 
-    def normal_mode(self):
-        viewmodel.global_mode.mode = Mode.NORMAL
+    def enter_navigate_mode(self):
+        viewmodel.global_mode.mode = Mode.NAVIGATE
 
     def trigger_filter(self):
         self.filter.active = True
@@ -186,7 +186,7 @@ class AssignControl(Control):
             return True
 
     def abort_assign_mode(self):
-        viewmodel.global_mode.mode = Mode.NORMAL
+        viewmodel.global_mode.mode = Mode.NAVIGATE
         self.list.render(self.list_size, True)
 
     def select_next(self):